% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_id_checksum_side}
\alias{all_id_checksum_side}
\title{Checksums of Participants}
\format{A vector of strings with id-checksum-side for the 160 participants.}
\source{
\url{http://http://forensicstats.org/}
}
\usage{
all_id_checksum_side
}
\description{
Checksums of Participants
}
\keyword{datasets}
