% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_files}
\alias{check_files}
\title{Check File Names}
\usage{
check_files(method, mode = NULL)
}
\arguments{
\item{method}{Integer specifying the method.}

\item{mode}{NULL by default. It specifies the mode to
check. Mandatory for method 1, optional for the rest.}
}
\description{
Check that every file in a direrctory matched the naming criterion
for a particular shoeprint method.
}
\author{
Guillermo Basulto Elias
}
