% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_individual_name}
\alias{check_individual_name}
\title{Check Shoeprint Name}
\usage{
check_individual_name(filename, checksum_table)
}
\arguments{
\item{filename}{Filename}

\item{checksum_table}{Table with checksums for ID and side.}
}
\value{
Logical value indicating if name is correct
}
\description{
Check if the name of a particular file is correct.
}
\author{
Guillermo Basulto-Elias
}
