% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Extract File Extension}
\usage{
get_file_extension(filename)
}
\arguments{
\item{filename}{File name}
}
\value{
character
}
\description{
Get the everything after the last dot in a filename
}
\author{
Guillermo Basulto-Elias
}
