% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_files_table}
\alias{print_files_table}
\title{Print tally of files}
\usage{
print_files_table(file_ext)
}
\arguments{
\item{file_ext}{Vector of characters with all the extensions founf
in a directory.}
}
\description{
Print tally of files
}
