% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_in_dir.R
\name{str_replace_in_dir}
\alias{str_replace_in_dir}
\title{Rename Files}
\usage{
str_replace_in_dir(pattern, replacement)
}
\arguments{
\item{pattern}{Pattern to be replaced. You may use regular
expressions.}

\item{replacement}{Replacement}
}
\description{
Look for a pattern in the files of a directoty and replace them
with an specified string.
}
\author{
Guillermo Basulto-Elias
}
